Forms top nav bar list and forms list
I was able to get rid of the yellow background the honors nav bar had on
hover when in rosponsive mode. Took away the yellow  background when
hovering over the list of forms Also when i first synced to master, i
opend the forms page and it had 2 honors nav bars. I deletled one of
them and updated the give today and honors application name.